fix(): test cases refactor and cleanup
