+++ pkg/schema/event_parser.go
